# docs/swagger.yaml
basePath: /
definitions:
...
  ukk-smkn2_features_kategorikarya_delete.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_kategorikarya_delete.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  ukk-smkn2_features_kategorikarya_getbyid.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_kategorikarya_getbyid.Response:
    properties:
      kategori_id:
        type: string
      nama:
        type: string
    type: object
  ukk-smkn2_features_kategorikarya_update.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_kategorikarya_update.Request:
    properties:
      nama:
        type: string
    type: object
  ukk-smkn2_features_kategorikarya_update.Response:
    properties:
      kategori_id:
        type: string
      nama:
        type: string
    type: object
  ukk-smkn2_features_review_create.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_review_create.Request:
    properties:
      karya_id:
        type: string
      komentar:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    type: object
  ukk-smkn2_features_review_create.Response:
    properties:
      karya_id:
        type: string
      komentar:
        type: string
      rating:
        type: integer
      review_id:
        type: string
      user_id:
        type: string
    type: object
  ukk-smkn2_features_review_delete.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_review_getall.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_review_getall.Response:
    properties:
      karya_id:
        type: string
      komentar:
        type: string
      rating:
        type: integer
      review_id:
        type: string
      user_id:
        type: string
    type: object
  ukk-smkn2_features_review_getbyid.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_review_getbyid.Response:
    properties:
      karya_id:
        type: string
      komentar:
        type: string
      rating:
        type: integer
      review_id:
        type: string
      user_id:
        type: string
    type: object
  ukk-smkn2_features_review_update.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_review_update.Request:
    properties:
      karya_id:
        type: string
      komentar:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    type: object
  ukk-smkn2_features_review_update.Response:
    properties:
      karya_id:
        type: string
      komentar:
        type: string
      rating:
        type: integer
      review_id:
        type: string
      user_id:
        type: string
    type: object
  ukk-smkn2_features_role_create.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_role_create.Request:
    properties:
      nama:
        type: string
    type: object
  ukk-smkn2_features_role_create.Response:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
...
  ukk-smkn2_features_role_getall.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_role_getall.Response:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: Restful API using go fiber
  title: your own API application
  version: "1.0"
paths:
...
  /api/kategori-karya/{id}:
    delete:
      description: Hapus satu kategori karya berdasarkan ID
      parameters:
      - description: ID Kategori Karya
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_delete.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_delete.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_delete.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_delete.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus Kategori Karya
      tags:
      - kategori_karya
    get:
      description: Mengembalikan detail kategori karya berdasarkan UUID
      parameters:
      - description: Kategori ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_getbyid.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_getbyid.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_getbyid.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil kategori karya berdasarkan ID
      tags:
      - kategori_karya
    put:
      consumes:
      - application/json
      description: Mengubah nama kategori karya berdasarkan ID
      parameters:
      - description: Kategori ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_kategorikarya_update.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_update.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_update.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_update.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_kategorikarya_update.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update kategori karya
      tags:
      - kategori_karya
  /api/review:
    get:
      description: Mengambil semua review yang ada di database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ukk-smkn2_features_review_getall.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_getall.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil Semua Review
      tags:
      - review
    post:
      consumes:
      - application/json
      description: Membuat data review untuk karya tertentu
      parameters:
      - description: Request body review
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_review_create.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_create.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_create.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_create.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat review baru
      tags:
      - review
  /api/review/{id}:
    delete:
      description: Hapus review berdasarkan ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_delete.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_delete.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_delete.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus Review
      tags:
      - review
    get:
      description: Mengambil satu data review berdasarkan ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_getbyid.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_getbyid.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_getbyid.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil Review by ID
      tags:
      - review
    put:
      consumes:
      - application/json
      description: Mengupdate data review berdasarkan ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Data review yang akan diupdate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_review_update.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_update.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_update.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_update.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_review_update.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Review by ID
      tags:
      - review
  /api/role:
    get:
      description: Get list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ukk-smkn2_features_role_getall.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_getall.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Roles
      tags:
      - role
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Role create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_role_create.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_create.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_create.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_create.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Role
      tags:
      - role
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"